Fix ILM rollover alias creation when a placeholder is used in index_name
This logic incorrectly assumes that if a template with the given template_name was previously
installed then it can safely do an early exit, even if the rollover alias was not created.

This corrects the logic so that it only does an early exit if both the rollover alias was created
AND the template was installed.